Handle possible date values in parsing term and termset queries
