+++ pg_search/src/query/pdb_query.rs
@@ -56,6 +56,7 @@ pub mod pdb {
         deserialize_bound, deserialize_bound_date_aware, serialize_bound,
         serialize_bound_date_aware,
     };
+    use crate::query::{deserialize_date_aware_owned_value, serialize_date_aware_owned_value};
     use pgrx::PostgresType;
     use serde::{Deserialize, Serialize};
     use std::collections::Bound;
@@ -316,6 +317,10 @@ pub mod pdb {
             max_expansions: Option<u32>,
         },
         Term {
+            #[serde(
+                serialize_with = "serialize_date_aware_owned_value",
+                deserialize_with = "deserialize_date_aware_owned_value"
+            )]
             value: OwnedValue,
         },
         TermSet {
